requirements.txt
langchain>=0.0.267
langchain-google-genai>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.0.0 